requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
textblob>=0.17.1
langdetect>=1.0.9